+++ project/media_items/models.py
@@ -11,9 +11,9 @@ from base.models import BaseModel
 
 @deconstructible
 class MediaItemPathValidator(validators.RegexValidator):
-    regex = r'^\/[0-9]{4}/[0-9]{4}-[0-9]{2}-[0-9]{2}/[0-9]{4}-[0-9]{2}-[0-9]{2}T[0-9]{6}_[0-9a-z]{8}\.[a-z]{3}$'
+    regex = r'^\/[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}/[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}\.[a-z]{3}$'
     message = _(
-        'Enter a valid media item path. Should be like this: `/yyyy/yyyy-mm-dd/yyyy-mm-ddThhmmss_slug.ext`'
+        'Enter a valid media item path. Should be like this: `/uuid/uuid.ext`'
     )
     flags = 0
 
@@ -69,19 +69,19 @@ class MediaItem(BaseModel):
     file_size = models.BigIntegerField(
         _('file size'),
         help_text=_('Required. The size in bytes of the media item.'),
-        null=True
+        null=False
     )
 
     image_width = models.IntegerField(
         _('image width'),
         help_text=_('Required. The width of this media item.'),
-        null=True
+        null=False
     )
 
     image_height = models.IntegerField(
         _('image height'),
         help_text=_('Required. The height of this media item.'),
-        null=True
+        null=False
     )
 
     camera_make = models.CharField(
@@ -180,17 +180,8 @@ class MediaItem(BaseModel):
         max_length=64
     )
 
-    def validate_unique(self, exclude=None):
-        o = MediaItem.objects.filter(path=self.file_path, owner=self.owner)
-        if o.exists():
-            raise validators.ValidationError({'path': _(
-                'There already exists a media item named [%s] for user [%s]' % (self.file_path, self.owner.username))})
-
     def __str__(self):
         return self.file_path
 
     class Meta:
         db_table = 'media_item'
-        unique_together = (
-            ('file_path', 'owner'),
-        )