+++ project/media_items/tests/test_upload_processing_integration.py
@@ -4,7 +4,6 @@ import subprocess
 from shutil import copyfile
 from glob import glob
 from urllib.parse import urlparse
-from uuid import UUID
 from datetime import datetime
 
 import pytest
@@ -12,6 +11,7 @@ from assertpy import assert_that
 
 from users.tests.factories import USER_PASSWORD
 from media_items.models import MediaItem
+from base.tests.util import match_image_key
 
 
 @pytest.mark.django_db
@@ -19,9 +19,8 @@ def test_sign_upload_request_success(authenticated_client, img, env):
     with(env['remote_path']):
         c, u = authenticated_client
 
-        media_id, image_key = request_upload(c, img)
-
-        image_key = '/%s%s' % (u.id, image_key)
+        image_key = request_upload(c, img)
+        media_id = parse_id_from_key(u.id, image_key)
 
         remote_file = mock_upload(img['local_path'], env['remote_path'], image_key)
         assert_that(remote_file).exists()
@@ -29,6 +28,14 @@ def test_sign_upload_request_success(authenticated_client, img, env):
         invoke_processor(image_key)
         actual = MediaItem.objects.get(id=media_id)
         assert_processing(img, actual)
+
+
+def parse_id_from_key(user_id, key):
+    m = match_image_key(str(user_id), key)
+    assert m is not None
+    id = m.group('image_id')
+    assert id is not None
+    return id
 
 
 def assert_processing(e, a):
@@ -69,13 +76,10 @@ def to_date(s):
 
 
 def request_upload(client, img):
-    response = client.post('/media/request-upload/', {'create_date': img['create_date'], 'mime_type': img['mime_type']})
+    response = client.post('/media/request-upload/', {'mime_type': img['mime_type']})
     assert response.status_code == 201
-    assert response['X-Elektron-Media-Id'] is not None
-    assert response['X-Elektron-Filename'] is not None
-    media_id = UUID(response['X-Elektron-Media-Id'])
-    path = response['X-Elektron-Filename']
-    return media_id, path
+    path = response['Location']
+    return urlparse(path).path
 
 
 def mock_upload(local, remote_path, image_key):
@@ -101,8 +105,8 @@ def run_processor(path):
     cwd = processor_project_dir()
     jar = '%s/build/libs/elektron-processor*.jar' % cwd
     jar = next(iter(glob(jar)), None)
-    if not jar:
-        raise FileNotFoundError('jarfile not found')
+    assert_that(jar).is_not_none()
+    assert_that(jar).exists()
     cmd = [
         'java',
         '-Dlog4j.configurationFile=log4j.properties',